backend/requirements.txt
fastapi
uvicorn
pymongo
certifi
python-dotenv
PyPDF2
mistralai
dnspython
requests